+++ packages/profiler-core/src/helper/LoggerHelper.ts
@@ -15,6 +15,10 @@ export class LoggerHelper {
 		return cli.red(util.format.apply(this, [message, ...optionalParams]))
 	}
 
+	static successString(message?: any, ...optionalParams: any[]) {
+		return cli.green(util.format.apply(this, [message, ...optionalParams]))
+	}
+
 	static log(message?: any, ...optionalParams: any[]) {
 		console.log(LoggerHelper.logString(message, ...optionalParams))
 	}
@@ -25,5 +29,9 @@ export class LoggerHelper {
 
 	static error(message?: any, ...optionalParams: any[]) {
 		console.error(LoggerHelper.errorString(message, ...optionalParams))
+	}
+
+	static success(message?: any, ...optionalParams: any[]) {
+		console.log(LoggerHelper.successString(message, ...optionalParams))
 	}
 }