+++ packages/profiler-core/src/types/helper/InspectorHelper.ts
@@ -1,0 +1,6 @@
+import { ISourceMap } from '../../model/SourceMap'
+
+export type IInspectorHelper = {
+	sourceCodeMap: Record<string, string>
+	sourceMapMap: Record<string, ISourceMap | null>
+}